agents/autogen-playwright/requirements.txt
pyautogen>=0.2.0
playwright>=1.39.0
python-dotenv>=0.19.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0 